+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SearchableList`:
   - Do initial search if `externalSearchTerm` is given
   - Correctly call expand/collapse function for heading
+- `Utils.htmlentities` and `Utils.htmlentities_decode`: Always return a string
 
 ## [2.0.0-rc.1] - 2020-01-22